feat: add details
