+++ src/components/PokemonCard.tsx
@@ -1,14 +1,8 @@
 import Link from 'next/link';
-import { pokemonAPI } from '../services/pokemonService';
-import Loading from './Loading';
 import { useRouter } from 'next/router';
+import { Pokemon } from '../api/types';
 
-export default function PokemonCard({ pokemonName }: { pokemonName: string }) {
-  const { isLoading, data: pokemon } =
-    pokemonAPI.useGetPokemonByNameQuery(pokemonName);
-  const { data: desc } = pokemonAPI.useGetPokemonDetailsQuery(
-    pokemon ? pokemon.name : ''
-  );
+export default function PokemonCard({ pokemon }: { pokemon: Pokemon }) {
   const router = useRouter();
 
   const url = `${router.basePath}/details/${pokemon?.id}`;
@@ -16,37 +10,32 @@ export default function PokemonCard({ pokemonName }: { pokemonName: string }) {
     <>
       <Link href={url}>
         <div className="card" data-testid="card">
-          {isLoading ? (
-            <Loading />
-          ) : (
-            <>
-              <h2>{pokemon?.name.toUpperCase()}</h2>
-              <h3>
-                Type:
-                {pokemon?.types.map((type) => (
-                  <span className="type" key={type.type.name}>
-                    {type.type.name}
-                  </span>
-                ))}
-              </h3>
-              <h4 className="size">
-                Height: {pokemon ? pokemon.height / 10 : 1} m
-              </h4>
-              <h4 className="size">
-                Weight: {pokemon ? pokemon.weight / 10 : 1} kg
-              </h4>
-              <img
-                src={
-                  pokemon?.sprites.front_default
-                    ? pokemon?.sprites.front_default
-                    : '/notAvaliable.jpg'
-                }
-                alt="pokemon"
-                width={150}
-              />
-              <p>{desc}</p>
-            </>
-          )}
+          <>
+            <h2>{pokemon?.name.toUpperCase()}</h2>
+            <h3>
+              Type:
+              {pokemon?.types.map((type) => (
+                <span className="type" key={type.type.name}>
+                  {type.type.name}
+                </span>
+              ))}
+            </h3>
+            <h4 className="size">
+              Height: {pokemon ? pokemon.height / 10 : 1} m
+            </h4>
+            <h4 className="size">
+              Weight: {pokemon ? pokemon.weight / 10 : 1} kg
+            </h4>
+            <img
+              src={
+                pokemon?.sprites.front_default
+                  ? pokemon?.sprites.front_default
+                  : '/notAvaliable.jpg'
+              }
+              alt="pokemon"
+              width={150}
+            />
+          </>
         </div>
       </Link>
     </>
