--- src/main.tsx
@@ -1,18 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-import { createBrowserRouter, RouterProvider } from 'react-router-dom';
-import './index.css';
-import { routerConfig } from './routerConfig';
-import { Provider } from 'react-redux';
-import { setupStore } from './store/store';
-
-const router = createBrowserRouter(routerConfig);
-const store = setupStore();
-
-ReactDOM.createRoot(document.getElementById('root')!).render(
-  <React.StrictMode>
-    <Provider store={store}>
-      <RouterProvider router={router} />
-    </Provider>
-  </React.StrictMode>
-);
