+++ src/store/store.ts
@@ -1,20 +1,16 @@
 import { combineReducers, configureStore } from '@reduxjs/toolkit';
 import searchReducer from './reducers/searchSlice';
 import perPageSlice from './reducers/perPageSlice';
-import { pokemonAPI } from '../services/pokemonService';
 import { createWrapper } from 'next-redux-wrapper';
 
 const rootReducer = combineReducers({
   searchReducer,
   perPageSlice,
-  [pokemonAPI.reducerPath]: pokemonAPI.reducer,
 });
 
 export const setupStore = () => {
   return configureStore({
     reducer: rootReducer,
-    middleware: (getDefaultMiddleware) =>
-      getDefaultMiddleware().concat(pokemonAPI.middleware),
   });
 };
